pyproject.toml
[project]
name = "pyex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
]
